works with green avatars now
found it weird that all the objects rendered by the camera into the rendertexture are actually transparent. However, since the cam.backgroundColor doesn't seem to be affected by the weird transparency, im instead setting the background as completely opaque (in contrast to the camera rendering objects as transparent), and the shader is instead removing the opaque pixels, leaving the "transparent" pixels of the avatar to be rendered as opaque... since its no longer using a "greenscreen" to replace it there probably shouldn't be any weird glow/work with avatars that contain green.