fixed alpha values
by inverting the alpha value in the shader (so that it is actually correct now), it is possible to use alphablending rather than simply discarding transparent pixels.
Currently, anti-aliasing is disabled for transparent background, this is because the anti-aliasing is done too early within the render texture with the broken alpha values. Trying to do the anti-aliasing within the shader instead for transparent.