Force AA to be disabled for avatarOnly
RenderTexture's AA is computed without taking background into account, resulting in white border around character when enabled. A possible solution would be to do the AA within the shader, when it is possible to access the background. Given the difficulty of doing so, AA is disabled for avatarOnly mode for now.